Avoid using xrandr for wine config if possible
